requirements.txt
alembic==1.7.7
click==8.0.4
colorama==0.4.4
datadispatch==1.0.0
Flask==2.0.2
Flask-Caching==1.10.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
importlib-resources==5.4.0
itsdangerous==2.1.1
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2==2.9.3
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.32
tasks==2.8.0
Werkzeug==2.0.3
zipp==3.7.0
